Rename from SimplifyIfElseReturnsRector to SimplifyReturnsRector
The new name describes better the Rector
