+++ config/level/code-quality/code-quality.yml
@@ -1,4 +1,4 @@
 services:
     Rector\Rector\CodeQuality\InArrayAndArrayKeysToArrayKeyExistsRector: ~
     Rector\Rector\CodeQuality\UnnecessaryTernaryExpressionRector: ~
-    Rector\Rector\CodeQuality\SimplifyIfElseReturnsRector: ~
+    Rector\Rector\CodeQuality\SimplifyReturnsRector: ~
