--- tests/Rector/CodeQuality/SimplifyIfElseReturnsRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\CodeQuality\SimplifyIfElseReturnsRector: ~
