+++ tests/Rector/CodeQuality/SimplifyReturnsRector/Correct/correct.php.inc
