+++ tests/Rector/CodeQuality/SimplifyReturnsRector/Wrong/wrong.php.inc
