+++ tests/Rector/CodeQuality/SimplifyReturnsRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\Rector\CodeQuality\SimplifyReturnsRector: ~
